wifi: fix connect event and boot DHCP on ESP32-C3
The ESP32 WiFi driver gates NET_EVENT_WIFI_CONNECT_RESULT on DHCP
completion when CONFIG_ESP32_WIFI_STA_AUTO_DHCPV4=y (the default).
WallaBMC also manages DHCP independently in net_init()/dhcp.c,
creating a double-start conflict that either prevents DHCP from
completing or blocks the connect event from ever firing.

Disable driver-managed DHCP (AUTO_DHCPV4=n) so the connect event
fires immediately on WiFi association, and let WallaBMC handle DHCP
exclusively.  Restart DHCP in the connect handler to get a fresh
lease promptly, and stop it on disconnect to avoid retransmitting
into the void.

Also disable CONFIG_ESP32_WIFI_STA_RECONNECT explicitly — WallaBMC
has its own reconnect logic with a 5-second backoff.

Fix net_mgmt callback signatures (uint32_t → uint64_t) to match the
Zephyr net_mgmt_event_handler_t typedef.  The mismatch corrupted the
iface parameter on 32-bit RISC-V due to ABI register layout, which
would crash now that the handlers use iface for DHCP calls.

+++ src/wifi_provision.c
@@ -51,7 +51,7 @@ static bool scan_in_progress;
 static struct net_mgmt_event_callback scan_cb;
 
 static void scan_result_handler(struct net_mgmt_event_callback *cb,
-				uint32_t mgmt_event, struct net_if *iface)
+				uint64_t mgmt_event, struct net_if *iface)
 {
 	if (mgmt_event == NET_EVENT_WIFI_SCAN_RESULT) {
 		const struct wifi_scan_result *entry =
